requirements.txt
flask
anthropic
requests
python-dotenv
schedule
cryptography
pytest
pytest-cov
